Add move and copy to daemon
