+++ daemon/service/new_session.go
@@ -8,7 +8,3 @@ import (
 func (daemon *FileSystemDaemon) NewSession(ctx context.Context, newSessionRequest *pb.NewSessionRequest) (*pb.NewSessionResponse, error) {
 	return daemon.sessionStore.AddSession(newSessionRequest, daemon.fs.DefaultWorkingDirectory())
 }
-
-func (daemon *FileSystemDaemon) DeleteSession(ctx context.Context, deleteSessionRequest *pb.DeleteSessionRequest) (*pb.DeleteSessionResponse, error) {
-	return daemon.sessionStore.DeleteSession(deleteSessionRequest)
-}
