+++ filesystem/filesystem.go
@@ -26,6 +26,7 @@ type FileSystem interface {
 	FindFiles(name string, path *fspath.FileSystemPath, workingDir file.File) ([]file.FileInfo, error)
 	ListFiles(path *fspath.FileSystemPath, workingDir file.File) ([]file.FileInfo, error)
 	Move(srcPath *fspath.FileSystemPath, destPath *fspath.FileSystemPath, workingDir file.File) (file.FileInfo, error)
+	Copy(srcPath *fspath.FileSystemPath, destPath *fspath.FileSystemPath, workingDir file.File) (file.FileInfo, error)
 }
 
 func NewFileSystem(fsType FileSystemType) (FileSystem, error) {
